pyproject.toml
[project]
name = "files"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ddgs>=9.13.0",
    "duckduckgo-search>=8.1.1",
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
